borders succesful
Added the logic to countryBordershandler to get the geojson feature of the country selected by the user. Updated the logic of the secces function of the ajax request to display the country and highlight it, using map.fitbounds to center it.

added a clear map button, still undecided if will be used.