add restCoutries.php
created the routine restcountries.php to make the request to the RESTCountries API and get information from country.
Declared global variables for currency, language, isocode, full name, etc that will be used to display data to the user in the future.
Remove the assigment of values to these variables from the call to ajaxOpenCage function as problaby theforward call to openCage will be obselete in the future, instead assigned the values using the json returned from RESTcountries API