covid19api added
Added an ajax request to coid19api, not using a php routine as this was not working due to a 404 not found error.
got the latest data on covid for the country. The json file actually contains data from day 1 and all days subsequentially so i used [length -1] to access the latest day.