HDR api added
Added the human development data.
replaced all the global variables that stored the data about the country in one new object: country. all the data is stored in the properties of the object now.
the weather data is not yet in the object, as not sure about how the data will be used