populate select added
Added logic to populate the select element using the countryGeoJson features. So now all the countries in the list are working.

Added logic to update the languages array of the country object to get 3 languages from the country.

Almost ready to layout phase but still no progresss trying to fix the problem where the onMapClick event doesnt highlighted the country on the first click...