BacktoMyLocation and weather object
Added functionality to back to my location buttom, it checks if the user is not already highlighting the country he is in.
Added the weather object and the logic to pass all data from the weather to this object. It grabs data from the location on the load of the page or with the Back to My Location buttom. Otherwise the user need to click on any place on the map to get info from openWeather.