Added place object
Added the place object, as well as the function placeUpdate that will be called by openCageUser and openCagereverse to get data from the user location or click event and populate the place object with useful information.
Adde dlogic to programatically send the wikipedia link from the country selected to the html countryWikipedia div.

That is the start of the layout phase of the project.