Address info added
Finished adding logic to display the country object at the map. Included a toggle info button so the user can hide it if suited.
Added a generatepopup function that is called on the mapclick event and when the page is loaded using user coords info. the popup shows timezone coordinates and address of the event.