Fixed a few bugs
Fixed bugs involving countries, changed ajaxRestCountries to use the iso code instead of the name of the country to solve a lot of bugs.
Identified a lot of countries with errors due to the HDI data API. Need to be addressed. All the countries listed in a .txt file called country issues.