Removed bugged countries and fixed others
Add logic to show 'not available ' for some countries that doesnt have data on Development indexes and removed bugged countries using '$(#countryOptions > option[value=-99]).remove()'.
Flags are buggy for a lot of countries.
Had to comment out the call to openExchangeRate due to reaching maximum free account allowance for the month, generalData() call transferred to ajaxrestCountries due to that reason.