wikipediaBoundinBox added
Added logic to gather wi,kipedia links for the bounding box of the country. Had to import leaflet-pip to the project via cdn to check if the wikipedia marker was located inside the geojson layer of the country because the bounding box is a rectangle and not exactly the country borders.