start to add weatherbtn logic
worked a bit on the popup markers logic for the wikipedia boundingBox ajax call.

Started to add logic to the weather btn, for now it is getting the geoname children for the country. Will have to further on call openweather for each of this children and displaya  different icon depending on the weather conditions. aAs well as the max min temp, wind speed and humidity.